+++ fabric/src/main/java/com/cozary/nameless_trinkets/events/MissingPageEvents.java
@@ -8,7 +8,9 @@ public class MissingPageEvents {
     public static void register() {
         ModEvents.DamageModifyCallback.EVENT.register((targetEntity, damageSource, amount) -> {
             if (targetEntity instanceof Player player) {
-                MissingPageHandler.triggerDamageReflection(player, damageSource, targetEntity);
+                if (amount > 0) {
+                    MissingPageHandler.triggerDamageReflection(player, damageSource, targetEntity);
+                }
             }
             return amount;
         });