+++ common/CHANGELOG.md
@@ -6,4 +6,5 @@
 
 ### Fixed
 
-- [Rage Mind] Fixed a server crash when an entity was damaged by a projectile. The game was incorrectly trying to identify the attacker, causing an error.
+- Fixed a server crash when an entity was damaged by a projectile. The game was incorrectly trying to identify the attacker, causing an error.
+- Fixed an issue where zero-damage events would incorrectly trigger the effects of several trinkets (e.g., Wooden Stick, Resonant Heart, Sigil of Baphomet).